FEND-946: Import templates into webapp
This is an import of some PLOS templates into the base Ambra webapp. It is
intended to bring the open-source Ambra front end in line with the
post-Ginsu back end and to minimize the number of custom templates that an
open-source user would need in order to host a journal. The imported
templates have had as many PLOS-specific elements as possible removed.

AmbraFreemarkerConfig now causes the complete set of CSS and JavaScript
files (with exceptions; see below) to be supplied for every page request if
not overridden in a journal.xml config file. The former default values were
obsolete and are always overridden in PLOS production. These default sets
deliberate exclude ga.js, because it is mostly PLOS-specific, and
highcharts.js, because its licensing permits it to be served to the user
only in non-commercial applications.

As of this revision, it should be possible for an installation with no
templates directory and no journal.xml files to render a (skeletal) front
page and article (if one is ingested).

Open issues: Many features haven't been checked or debugged yet. The
JavaScript code still contains hard-coded PLOS branding (appearing on
lightboxes, floating banner, etc.). The CSS needs to be changed to a new,
"generic Ambra" design rather than resembling PLOS journals. Miscellaneous
glitches with CSS and FreeMarker values.