Turn off DTD validation during parsing of article XML for display.
The package org.ambraproject.xml.transform is intended to serve XML entites (such as DTDs) from files on the classpath.  However, using ngrep on my dev box, I saw that it is not doing this completely, and we are still making HTTP calls to nlm.nih.gov to retrieve these during the article transform (29 such requests!).  I suspect that this package has not been kept up-to-date when the NLM DTD has changed.  This changeset provides a mechanism to completely bypass the loading of a DTD and all its associated entities, which is what we want at article display time since the XML is assumed to be valid.