Modify JSON method so that article counts of categories are returned along with the category names.
For simplicity, and to maintain backwards compatibility with anything that is consuming the current category JSON (which I think includes the indexer minion), I added a new dict "counts" to the output instead of decorating the categories directly with the counts.

I think this code is ripe for a complete refactor, but since we're launching this feature soon (with the new PLOS ONE homepage), I opted to work within the existing design.

TODO: "roll up" article counts recursively for non-leaf categories.