Implement article counts.  Delete code related to my earlier, erroneous attempts at this.
In order to match the number of search results on browse pages, I had to issue a new solr query that looks at the subject_facet facet instead of subject_hierarchy, which is used to populate the taxonomy browser.  However the results of this query are cached for 24 hours, just like the "main" taxonomy browser query.